Arrows!
About arrows, the main idea here is that they get colored by the complementary color of the button it takes place in.
This actually looks like pretty simple to code, maybe I can do this if prefered.

But it isn't in any way a priority, only a wild idea. The arrow should work beautifully with any color.