Arrow & Quality of life .gitgnore improvement
- The full "Loop Arrow"
- The separate parts of "Loop Arrow"
- The Unity Object of the separate parts of "Loop Arrow"

And now, allow me to explain the "quality-of-life improvement"
Every time I opened Github, it asked me if I wanted to commit the folder which is a mess and filled with things that doesn't matter to the end project (mostly .ase files and drafts). It also asked me if I wanted to commit the Palhaço Gozo's concept I've made.
I've added them all (the folder and the Palhaço Gozo's concept) to .gitgnore so I wouldn't accidentally commit them in a needlessly large and futile commit. It doesn't improve much on you, but it's way less maddening to me to deal with the github client.

It's futile, but at least is easily reversible if something goes terribly wrong.